Integrate killerUpdate and elementalAssistsUpdate commands with the corresponding infographic show commands Improve documentation of the infographic generation Remove debug messages Remove some unused code
